Fix Code after moving Contact into Models Package
