+++ app/build.gradle
@@ -13,6 +13,7 @@ android {
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+
     }
 
     buildTypes {
@@ -55,4 +56,8 @@ dependencies {
     implementation 'com.google.android.material:material:1.2.1'
     //Json
     implementation 'com.codepath.libraries:asynchttpclient:2.2.0'
+    implementation 'com.squareup.okhttp3:okhttp:4.1.0'
+    implementation 'com.fasterxml.jackson.core:jackson-databind:2.8.5'
+    implementation 'com.fasterxml.jackson.core:jackson-core:2.8.5'
+    implementation 'com.fasterxml.jackson.core:jackson-annotations:2.8.5'
 }
