+++ app/src/main/java/com/example/assistgoandroid/Call/VideoCallRinging.java
@@ -12,8 +12,8 @@ import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.bumptech.glide.Glide;
-import com.example.assistgoandroid.Contact.Contact;
 import com.example.assistgoandroid.R;
+import com.example.assistgoandroid.models.Contact;
 
 import java.util.Calendar;
 import java.util.Date;
@@ -39,9 +39,9 @@ public class VideoCallRinging extends AppCompatActivity {
         contact = (Contact) getIntent().getParcelableExtra("CONTACT_CARD");
         Log.i(TAG, "Contact is " + contact);
 
-        contactName.setText(contact.getName());
+        contactName.setText(contact.getFullName());
         Glide.with(this)
-                .load(contact.getContactPicture())
+                .load(contact.getProfileImageUrl())
                 .centerCrop()
                 .fitCenter() // scale to fit entire image within ImageView
                 .transform(new RoundedCornersTransformation(500,10))
