+++ app/src/main/java/com/example/assistgoandroid/Contact/newContactCardActivity.java
@@ -65,6 +65,7 @@ public class newContactCardActivity extends AppCompatActivity {
                 checkPermissionForNewContact();
             else
                 addNewContact();
+            //todo sync contact list with backend
         }
     }
